+++ src/main/java/com/alex/adventCode/DayOne.java
@@ -23,7 +23,7 @@ public class DayOne {
 	private static final Pattern VALUE_PATTERN = Pattern.compile("^(\\+|-){1}(\\d+){1}$");
 
 	public long getFrequency(File input) throws FileNotFoundException {
-		Long currentFrequency = 0L;
+		Integer currentFrequency = 0;
 		try (Scanner scanner = new Scanner(input)) {
 			while (scanner.hasNextLine()) {
 				String line = scanner.nextLine();
@@ -33,18 +33,18 @@ public class DayOne {
 		return currentFrequency;
 	}
 
-	private long updateFrequency(long frequency, String line) {
+	private Integer updateFrequency(Integer frequency, String line) {
 		Matcher matcher = VALUE_PATTERN.matcher(line);
 		Validate.isTrue(matcher.matches());
 		if (matcher.matches()) {
-			long value = Long.parseLong(matcher.group(VALUE_INDEX));
+			Integer value = Integer.parseInt(matcher.group(VALUE_INDEX));
 			String operator = matcher.group(OPERATOR_INDEX);
 			return addValue(frequency, value, operator);
 		}
 		return frequency;
 	}
 
-	private long addValue(long frequency, long value, String operator) {
+	private Integer addValue(Integer frequency, Integer value, String operator) {
 		if (operator.equals(ADDITION_OPERATOR)) {
 			frequency = frequency + value;
 		} else if (operator.equals(SUBTRACTION_OPERATOR)) {
@@ -54,8 +54,8 @@ public class DayOne {
 	}
 
 	public long findFirstDuplicateFrequency(File input) throws Exception {
-		Long currentFrequency = 0L;
-		Set<Long> frequencies = new HashSet<>();
+		Integer currentFrequency = 0;
+		Set<Integer> frequencies = new HashSet<>();
 		boolean hasFoundDuplicate = false;
 		List<String> driftFrequencies = Files.lines(input.toPath()).collect(toList());
 
@@ -67,8 +67,8 @@ public class DayOne {
 		return currentFrequency;
 	}
 
-	private FrequencyResult parseFrequency(List<String> driftFrequencies, Set<Long> frequencies, Long currentFrequency)
-			throws Exception {
+	private FrequencyResult parseFrequency(List<String> driftFrequencies, Set<Integer> frequencies,
+			Integer currentFrequency) throws Exception {
 		for (String line : driftFrequencies) {
 			currentFrequency = updateFrequency(currentFrequency, line);
 
@@ -84,9 +84,9 @@ public class DayOne {
 
 	private static class FrequencyResult {
 		private final boolean foundDuplicate;
-		private final Long frequency;
+		private final Integer frequency;
 
-		public FrequencyResult(boolean foundDuplicate, Long frequency) {
+		public FrequencyResult(boolean foundDuplicate, int frequency) {
 			this.foundDuplicate = foundDuplicate;
 			this.frequency = frequency;
 		}
